Crate a new external_api module
