+++ src/controller_examples/zookeeper_controller/exec/zookeeper_lib/lib.rs
@@ -1,7 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
+use crate::external_api::exec::*;
 use crate::pervasive_ext::to_view::*;
-use crate::reconciler::exec::external::*;
 use crate::zookeeper_controller::common::*;
 use crate::zookeeper_controller::exec::zookeeper_lib::helper_funcs::*;
 use crate::zookeeper_controller::exec::zookeepercluster::*;
