+++ src/external_api/mod.rs
@@ -1,4 +1,4 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
-pub mod io;
-pub mod reconciler;
+pub mod exec;
+pub mod spec;
