+++ src/external_api/spec.rs
@@ -1,0 +1,38 @@
+// Copyright 2022 VMware, Inc.
+// SPDX-License-Identifier: MIT
+use crate::pervasive_ext::to_view::*;
+use vstd::prelude::*;
+
+verus! {
+
+pub trait ExternalAPI {
+
+    type Input;
+    type Output;
+    type State;
+
+    open spec fn transition(input: Self::Input, state: Self::State) -> (Option<Self::Output>, Self::State);
+
+    open spec fn init_state() -> Self::State;
+}
+
+pub struct EmptyType {}
+
+pub struct EmptyAPI {}
+
+impl ExternalAPI for EmptyAPI {
+
+    type Input = EmptyType;
+    type Output = EmptyType;
+    type State = EmptyType;
+
+    open spec fn transition(input: EmptyType, state: EmptyType) -> (Option<EmptyType>, EmptyType) {
+        (Option::None, EmptyType{})
+    }
+
+    open spec fn init_state() -> EmptyType {
+        EmptyType {}
+    }
+}
+
+}
