+++ src/fluent_controller.rs
@@ -7,6 +7,7 @@ pub mod kubernetes_api_objects;
 pub mod kubernetes_cluster;
 pub mod pervasive_ext;
 pub mod reconciler;
+pub mod external_api;
 pub mod shim_layer;
 pub mod state_machine;
 pub mod temporal_logic;
