+++ src/kubernetes_cluster/spec/controller/common.rs
@@ -1,9 +1,9 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
+use crate::external_api::spec::*;
 use crate::kubernetes_api_objects::{api_method::*, common::*, resource::*};
-use crate::kubernetes_cluster::spec::external_api::*;
-use crate::kubernetes_cluster::spec::message::*;
+use crate::kubernetes_cluster::spec::{external_api::*, message::*};
 use crate::reconciler::spec::reconciler::*;
 use crate::state_machine::action::*;
 use crate::state_machine::state_machine::*;
