+++ src/kubernetes_cluster/spec/controller/controller_runtime.rs
@@ -1,6 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
+use crate::external_api::spec::*;
 use crate::kubernetes_api_objects::{api_method::*, common::*, resource::*};
 use crate::kubernetes_cluster::spec::{controller::common::*, external_api::*, message::*};
 use crate::reconciler::spec::{io::*, reconciler::*};
