+++ src/kubernetes_cluster/spec/external_api.rs
@@ -1,6 +1,7 @@
 // Copyright 2022 VMware, Inc.
 // SPDX-License-Identifier: MIT
 #![allow(unused_imports)]
+use crate::external_api::spec::*;
 use crate::kubernetes_cluster::spec::message::*;
 use crate::state_machine::action::*;
 use crate::state_machine::state_machine::*;
@@ -8,17 +9,6 @@ use crate::temporal_logic::defs::*;
 use vstd::{multiset::*, prelude::*};
 
 verus! {
-
-pub trait ExternalAPI {
-
-    type Input;
-    type Output;
-    type State;
-
-    open spec fn transition(input: Self::Input, state: Self::State) -> (Option<Self::Output>, Self::State);
-
-    open spec fn init_state() -> Self::State;
-}
 
 #[is_variant]
 pub enum ExternalComm<Input, Output> {
@@ -40,25 +30,6 @@ pub struct ExternalAPIState<T: ExternalAPI> {
 pub open spec fn external_api_state_init<T: ExternalAPI>(s: ExternalAPIState<T>) -> bool {
     &&& s.external_api_state == T::init_state()
     &&& s.in_flight == Set::<ExternalComm<T::Input, T::Output>>::empty()
-}
-
-pub struct EmptyType {}
-
-pub struct EmptyAPI {}
-
-impl ExternalAPI for EmptyAPI {
-
-    type Input = EmptyType;
-    type Output = EmptyType;
-    type State = EmptyType;
-
-    open spec fn transition(input: EmptyType, state: EmptyType) -> (Option<EmptyType>, EmptyType) {
-        (Option::None, EmptyType{})
-    }
-
-    open spec fn init_state() -> EmptyType {
-        EmptyType {}
-    }
 }
 
 }
